Genericly test overlay support on the storage directory
The old code only tested overlay-over-overlay. In any other case, it switched
to a loop mounted ext4 with overlay or aufs on-top.

This PR activates the smoke test independently from the storage filesystem.
Hence, we will use overlay natively on the /var/lib/docker also if e.g. it is
mounted from the host as a volume (for image caching purposes).